refactor: 콘텐츠 상세 바텀시트 패키지 이동
