+++ android/app/src/main/java/com/on/turip/ui/compose/trip/bottomsheet/PlaceTuripSelectionUiEffect.kt
@@ -1,7 +1,7 @@
-package com.on.turip.ui.compose.turip.selection
+package com.on.turip.ui.compose.trip.bottomsheet
 
 import com.on.turip.ui.common.error.ErrorUiState
-import com.on.turip.ui.compose.turip.selection.model.TuripPlaceModel
+import com.on.turip.ui.compose.trip.bottomsheet.model.TuripPlaceModel
 import com.on.turip.ui.main.favorite.model.TuripModel
 import com.on.turip.ui.main.favorite.model.TuripShareModel
 import kotlinx.collections.immutable.ImmutableList
