+++ android/app/src/main/java/com/on/turip/ui/compose/trip/bottomsheet/PlaceTuripSelectionUiState.kt
@@ -1,7 +1,7 @@
-package com.on.turip.ui.compose.turip.selection
+package com.on.turip.ui.compose.trip.bottomsheet
 
 import androidx.compose.runtime.Immutable
-import com.on.turip.ui.compose.turip.selection.model.TuripPlaceModel
+import com.on.turip.ui.compose.trip.bottomsheet.model.TuripPlaceModel
 import com.on.turip.ui.main.favorite.model.TuripModel
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
