+++ android/app/src/main/java/com/on/turip/ui/compose/trip/bottomsheet/model/TuripPlaceModel.kt
@@ -1,4 +1,4 @@
-package com.on.turip.ui.compose.turip.selection.model
+package com.on.turip.ui.compose.trip.bottomsheet.model
 
 import androidx.compose.runtime.Stable
 import com.google.android.gms.maps.model.LatLng
