+++ android/app/src/main/java/com/on/turip/ui/compose/trip/bottomsheet/util/reorderable/ReorderableLazyColumnState.kt
@@ -1,4 +1,4 @@
-package com.on.turip.ui.compose.turip.selection.util.reorderable
+package com.on.turip.ui.compose.trip.bottomsheet.util.reorderable
 
 import androidx.compose.animation.core.Animatable
 import androidx.compose.animation.core.Spring
@@ -15,6 +15,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.geometry.Offset
+import kotlin.coroutines.cancellation.CancellationException
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.take
@@ -22,7 +23,6 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 import kotlinx.coroutines.withTimeout
-import kotlin.coroutines.cancellation.CancellationException
 
 /**
  * 스크롤 가능한 순서 변경을 위해 정의한 State
