+++ android/app/src/main/java/com/on/turip/ui/main/favorite/PlaceTuripSelectionFragment.kt
@@ -21,8 +21,8 @@ import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.on.turip.R
 import com.on.turip.ui.common.extensions.safeStartActivityWithToast
 import com.on.turip.ui.compose.designsystem.theme.TuripTheme
-import com.on.turip.ui.compose.turip.selection.PlaceTuripSelectionBottomSheet
-import com.on.turip.ui.compose.turip.selection.PlaceTuripSelectionViewModel
+import com.on.turip.ui.compose.trip.bottomsheet.PlaceTuripSelectionBottomSheet
+import com.on.turip.ui.compose.trip.bottomsheet.PlaceTuripSelectionViewModel
 import com.on.turip.ui.folder.TuripActivity
 import com.on.turip.ui.login.LoginActivity
 import com.on.turip.ui.main.favorite.model.TuripShareModel
